pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-rl"
version = "0.1.0"
description = "OpenEnv ETL data remediation environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "firebase-admin>=6.5.0,<7.0.0",
    "google-genai>=1.0.0,<2.0.0",
    "numpy>=2.0.0,<3.0.0",
    "openenv-core>=0.2.0",
    "pandas>=2.2.0,<3.0.0",
    "pydantic>=2.8.0,<3.0.0",
    "uvicorn>=0.35.0,<1.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["app", "inference", "preload_models"]

[tool.setuptools.packages.find]
include = ["nova_rl_env", "server"]
